feat: use JSON cache for generating the entire view logic
This is just a lot of moved code. As we have the cache anyways, we only
save time without introducing further invalidation problems.
